Make `TimestampTemporal` internal again and separate `Instant` conversion.
